Services

+++ back/src/main/java/com/example/back/repository/GPURepository.java
@@ -3,5 +3,5 @@ package com.example.back.repository;
 import com.example.back.model.GPU;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GPURepository extends JpaRepository<GPU, Integer> {
+public interface GPURepository extends JpaRepository<GPU, Long> {
 }
